requirements.txt
fastapi~=0.104.1
uvicorn[standard]~=0.24.0.post1
python-multipart
fastapi-login
passlib[bcrypt]~=1.7.4
python-jose
jinja2
rich
pydantic~=2.5.2
environs~=9.5.0
asyncpg~=0.29.0
loguru~=0.7.2